pyproject.toml
[project]
name = "cheap_electricity"
version = "0.1.0"
description = ""
authors = [
    {name = "Gemini"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "python-telegram-bot (>=22.3,<23.0)",
    "nest-asyncio (>=1.6.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pandas (>=2.2.2,<3.0.0)"
]

[tool.poetry.scripts]
run_main = "cheap_electricity.main:run_main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

